requirements.txt
Django==1.8.3
honcho==0.6.6
psycopg2==2.6.1
django-taggit==0.16.2
django-haystack==2.4.0
git+https://github.com/notanumber/xapian-haystack.git
Pillow==2.9.0
celery==3.1.18
redis==2.10.3
django-concertina==1.0.0
biplist==0.9
beautifulsoup4==4.4.0
celery-haystack==0.9
gunicorn==19.3.0
watchdog==0.8.3
pyScss==1.3.4
Markdown==2.6.2
django-cape==1.0
django-render-as==1.0
